+++ src/decorator.ts
@@ -1,0 +1,112 @@
+import * as vscode from "vscode";
+import { getStepFileUri } from "./utils";
+import { store, CodeTour, CodeTourStep } from "./store";
+import { reaction } from "mobx";
+import { ICON_URL } from "./constants";
+
+const TOUR_DECORATOR = vscode.window.createTextEditorDecorationType({
+  gutterIconPath: vscode.Uri.parse(ICON_URL),
+  gutterIconSize: "contain",
+  overviewRulerColor: "rgb(246,232,154)",
+  overviewRulerLane: vscode.OverviewRulerLane.Right
+});
+
+type CodeTourStepTuple = [CodeTour, CodeTourStep, number];
+
+async function getTourSteps(
+  document: vscode.TextDocument,
+  lineNumber?: number
+): Promise<CodeTourStepTuple[]> {
+  const steps: CodeTourStepTuple[] = store.tours.flatMap(tour =>
+    tour.steps.map(
+      (step, stepNumber) => [tour, step, stepNumber] as CodeTourStepTuple
+    )
+  );
+
+  const tourSteps = await Promise.all(
+    steps.map(async ([tour, step, stepNumber]) => {
+      const workspaceRoot = vscode.workspace.workspaceFolders![0]!.uri.toString();
+      const uri = await getStepFileUri(step, workspaceRoot);
+
+      if (
+        uri.toString().localeCompare(document.uri.toString()) === 0 &&
+        (!lineNumber || step.line! - 1 === lineNumber)
+      ) {
+        return [tour, step, stepNumber];
+      }
+    })
+  );
+
+  // @ts-ignore
+  return tourSteps.filter(i => i);
+}
+
+async function setDecorations(editor: vscode.TextEditor) {
+  const tourSteps = await getTourSteps(editor.document);
+  if (tourSteps.length === 0) {
+    return;
+  }
+
+  const ranges = tourSteps.map(
+    ([, step]) => new vscode.Range(step.line! - 1, 0, step.line! - 1, 1000)
+  );
+
+  editor.setDecorations(TOUR_DECORATOR, ranges);
+}
+
+function registerHoverProvider() {
+  return vscode.languages.registerHoverProvider("*", {
+    provideHover: async (
+      document: vscode.TextDocument,
+      position: vscode.Position
+    ) => {
+      const tourSteps = await getTourSteps(document, position.line);
+      if (tourSteps.length === 0) {
+        return;
+      }
+
+      const hovers = tourSteps.map(([tour, _, stepNumber]) => {
+        const args = encodeURIComponent(JSON.stringify([tour.id, stepNumber]));
+        const command = `command:codetour._startTourById?${args}`;
+        return `CodeTour: ${tour.title} (Step #${stepNumber +
+          1})&nbsp;&nbsp;&nbsp;[Start Tour](${command} "Start Tour")`;
+      });
+
+      const content = new vscode.MarkdownString(hovers.join("<br />"));
+      content.isTrusted = true;
+      return new vscode.Hover(content);
+    }
+  });
+}
+
+let disposables: vscode.Disposable[] = [];
+export async function registerDecorators() {
+  reaction(
+    () => [
+      store.showMarkers,
+      store.tours.map(tour => [tour.title, tour.steps])
+    ],
+    () => {
+      const activeEditor = vscode.window.activeTextEditor;
+      if (store.showMarkers) {
+        disposables.push(registerHoverProvider());
+
+        disposables.push(
+          vscode.window.onDidChangeActiveTextEditor(editor => {
+            if (editor) {
+              setDecorations(editor);
+            }
+          })
+        );
+
+        if (activeEditor) {
+          setDecorations(activeEditor);
+        }
+      } else if (activeEditor) {
+        activeEditor.setDecorations(TOUR_DECORATOR, []);
+        disposables.forEach(disposable => disposable.dispose());
+        disposables = [];
+      }
+    }
+  );
+}