+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+## v0.0.17 (03/31/2020)
+
+- Introduced "tour markers", which display a gutter icon next to lines of code which are associated with a step in a code tour.
+
 ## v0.0.16 (03/30/2020)
 
 - Updated the `CodeTour` tree to display the currently active tour, regardless how it was started (e.g. you open a tour file).
-  
+
 ## v0.0.15 (03/29/2020)
 
 - Updated the `CodeTour` tree to only display if the currently open workspace has any tours, or if the user is currently taking a tour. That way, it isn't obtrusive to users that aren't currently using it.